+++ securedrop/source_templates/base.html
@@ -15,7 +15,7 @@
     <div class="content">
       {% block header %}
       <div id="header">
-        <a href="{% if 'logged_in' in session %}{{ url_for('lookup') }}{% else %}{{ url_for('main.index') }}{% endif %}">
+        <a href="{% if 'logged_in' in session %}{{ url_for('main.lookup') }}{% else %}{{ url_for('main.index') }}{% endif %}">
           <img src="/static/i/{{ header_image }}" class="logo small" alt="SecureDrop" width="250px">
         </a>
         {% include 'locales.html' %}
