+++ securedrop/source_templates/lookup.html
@@ -38,7 +38,7 @@
 
   <hr class="no-line">
   <div class="pull-right">
-    <a href="{{ url_for('lookup') }}" class="btn secondary" id="cancel">{{ gettext('CANCEL') }}</a>
+    <a href="{{ url_for('main.lookup') }}" class="btn secondary" id="cancel">{{ gettext('CANCEL') }}</a>
     <button type="submit" class="sd-button btn primary" id="submit-doc-button">
       <img class="icon off-hover" src="{{ url_for('static', filename='i/font-awesome/cloud-upload-white.png') }}" width="20px" height="14px">
       <img class="icon on-hover" src="{{ url_for('static', filename='i/font-awesome/cloud-upload-blue.png') }}" width="20px" height="14px">
